% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-download_gnr_data.R
\name{download_gnr_country_profiles}
\alias{download_gnr_country_profiles}
\title{Download GNR country profiles dataset}
\usage{
download_gnr_country_profiles(path = ".")
}
\arguments{
\item{path}{Directory to download the specified report into. Default to
current working directory}
}
\value{
A tibble  providing details of file downloaded.
}
\description{
Download GNR country profiles dataset
}
\examples{
download_gnr_country_profiles(path = tempdir())

}
