% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_gnr_data.R
\name{download_gnr_data}
\alias{download_gnr_data}
\title{Download GNR data}
\usage{
download_gnr_data(.year = as.character(c(2018, 2020, 2021, 2022)), path = ".")
}
\arguments{
\item{.year}{A character value of the report year to download. Default to
"2018"}

\item{path}{Directory to download the specified report into. Default to
current working directory}
}
\value{
A tibble  providing details of file downloaded.
}
\description{
Download GNR data
}
\examples{
download_gnr_data(path = tempdir())

}
