% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_gnr_data.R
\name{download_n4g_data}
\alias{download_n4g_data}
\title{Download Nutrition for Growth Assessments data}
\usage{
download_n4g_data(path = ".")
}
\arguments{
\item{path}{Directory to download the specified report into. Default to
current working directory}
}
\value{
A tibble  providing details of file downloaded.
}
\description{
Download Nutrition for Growth Assessments data
}
\examples{
download_n4g_data(path = tempdir())

}
