% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{genenetwork-data}
\alias{genenetwork-data}
\alias{gn_species}
\alias{gn_groups}
\alias{gn_datasets}
\alias{gn_genotypes}
\title{Retrieve Data from GeneNetwork}
\usage{
gn_species()

gn_groups(species = NULL)

gn_datasets(group)

gn_genotypes(group)
}
\arguments{
\item{species}{A valid species name (e.g., \code{"mouse"}). Use \code{gn_species()}
to retrieve a list of species supported by GeneNetwork.}

\item{group}{A group name. Use \code{gn_groups()} to retrieve a list of valid
group names.}
}
\description{
Retrieve Data from GeneNetwork
}
\section{Functions}{
\itemize{
\item \code{gn_species}: Table of supported species (e.g., \code{"mouse"})

\item \code{gn_groups}: Table of available groups of subjects or
samples (e.g., \code{"bxd"}).

\item \code{gn_datasets}: Table of available datasets for a particular \code{group}.

\item \code{gn_genotypes}: Retrieve genotypes data for a specific \code{group}.
}}

\examples{
species <- gn_species()
groups <- gn_groups()

# retrieve all bxd datasets
datasets <- gn_datasets(group = "bxd")

# retrieve bxd genotypes
genos <- gn_genotypes(group = "bxd")
}
