% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnr.R
\docType{package}
\name{gnr}
\alias{gnr}
\alias{gnr-package}
\title{gnr: R companion to the Global Nutrition Report}
\description{
The \code{gnr} package is a companion package to the Global Nutrition Report
primarily for accessing the data used in the report and to reproduce the data
visualisation produced in the report. This package contains functions that
allow users to interrogate the same data and produce customised charts and
other visualisation products.
}
