% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gnr_data}
\alias{gnr_data}
\title{Download links for various GNR datasets}
\format{
A tibble with 4 columns and 5 rows:

**Variable** | **Description**
:--- | :---
**year** | Year dataste was created
**code** | Dataset code
**name** | Name of dataset
**link** | Download link for dataset
}
\source{
[Global Nutrition Report](https://globalnutritionreport.org)
}
\usage{
gnr_data
}
\description{
Download links for various GNR datasets
}
\keyword{datasets}
