% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicator_description}
\alias{indicator_description}
\title{Description of variables that occurs in the Global Nutrition Report (GNR)
dataset}
\format{
A data.frame with 709 rows and 3 columns:
\describe{
\item{\code{section}}{Section of the report}
\item{\code{variable}}{Data variable}
\item{\code{description}}{Description of variable}
}
}
\source{
\href{https://globalnutritionreport.org}{Global Nutrition Report}
}
\usage{
indicator_description
}
\description{
Description of variables that occurs in the Global Nutrition Report (GNR)
dataset
}
\examples{
indicator_description

}
\keyword{datasets}
