% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-theme_gnr.R
\name{theme_gnr}
\alias{theme_gnr}
\title{Global Nutrition Report ggplot theme}
\usage{
theme_gnr(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{Base font size}

\item{base_family}{Base font family}
}
\value{
A ggplot theme
}
\description{
Global Nutrition Report ggplot theme
}
\examples{
# Apply theme
theme_gnr()

}
