% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gnr_data}
\alias{gnr_data}
\title{Download links for various GNR datasets}
\format{
A tibble with 5 columns and 6 rows:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \strong{year} \tab Year dataste was created \cr
   \strong{code} \tab Dataset code \cr
   \strong{file_type} \tab Download file type \cr
   \strong{name} \tab Name of dataset \cr
   \strong{link} \tab Download link for dataset \cr
}
}
\source{
\href{https://globalnutritionreport.org}{Global Nutrition Report}
}
\usage{
gnr_data
}
\description{
Download links for various GNR datasets
}
\keyword{datasets}
